Issue #164: Added two gauges to monitor the circuit breaker state.
* Added two gauges to monitor the circuit breaker state. One checks the
overall "open" state, the other one just reflects the state overall
(numerically).

The idea is to provide the ability to monitor on either of the two
metrics.

* Report the state of the CB as an individual metric

* Collapsing the state metrics to a single metric with an integer that
represents each state. The value is tied to the state enumeration.

* adding a fixed final int to the state so regardless of enum order the
integer representation states consistent. Made minor additions to the
dropwizard metrics doc