Issue #178: add validation to CircuitBreakerProperties
CircuitBreakerProperties:
- removed @Component (not necessary due to usage of
@EnableConfigurationProperties in CircuitBreakerAutoConfiguration)
- Spring-Boot 2 supports using java.time.Duration as Property type. So I
changed the type of waitDurationInOpenState. This also fixes a
difference between the javadoc and the code (javadoc said it would be
seconds, but it was milliseconds).
- added validation for properties. Application should fail fast on
mis-configuration. For this feature I had to add the hibernate-validator
dependency.
- changed default behaviour to registerHealthIndicator=true. Convention
over configuration - if you provide a feature, you should activate by
default.

CircuitBreakerMetricsAutoConfiguration:
- add possibility to disable registering of circuitbreaker metrics