Additional factory methods
These factory methods are identical to the ones already existing in
metrics/CircuitBreakerMetrics.java. Their purpose is to allow
Spring Boot applications to instantiate CircuitBreakerMetrics and
have fine-grained control over which circuit breakers are bound to
metrics, in particular in scenarios where the CircuitBreakerRegistry
may not be completely populated by the time the CircuitBreakerMetrics
class is instantiated.