Fix formatting of Bulkhead section in documentation
The extra blank line at the end of ratelimiter.adoc fixes the formatting for the Bulkhead section
(or any .adoc) included after it.
The end of ratelimiter.adoc was special because the other .adoc files in the same directory end
with a source block while ratelimiter.adoc ends with a simple paragraph.