Update Gradle to 5.3.1
This commit updates gradle to 5.3.1.
To make the update possible all plugins were updated to the latest version as well.
Also the jcstress plugin doesn't seem to be compatible with gradle 5.
Given that the last development activity for this plugin was 6 month ago,
I decided to remove jcstress related plugines completly so it doesn't prevent future updates.