Issue #509: Support for extending Micrometer tags
For extending the tags provided by resilience4j, the classes that implement
MeterBinder for the respective concern(Retry, CircuitBreaker, RateLimiter
and Bulkhead) need to be extendable. With this commit, they all now implement
respective interfaces and these interfaces are set as the return type for
the respective spring boot auto configuration.