Add support for Fegin 12.5+ Feign 12.5 made the builder.build() method final, and started internally clone it...
This breaks resilience4j integration with feign and renders resilience4j completely unused by feign.

In version 10.9 they introduced an api to add "capabilities" to the feign client.

Here I have made changed to the 1.7 branch (the release I need) that implements this new pattern instead, and provides a fallback using old pattern for backwards compatibility down to 10.9.
Cherry-pick to master note: As this version of resilience4j already runs feign 12.0, no changes to master in what feign version we build with