fix spring boot3 customizer application order (#2321)
* Feature: port changes to customizer run order from spring-boot2 to spring-boot3

* Fix: fix system-dependent SpringBootCommonTest test failures

After applying the changes to the customizer initialization order,
SpringBootCommonTest is now failing due to the test attempting to set
`coreThreadPoolSize` without setting the `maxThreadPoolSize`.

This was not triggering an error before because the properties were not
actually set as `backend` properties were not initialized and so the
customizer never actually ran.